fix(perf): simplify benchmark entrypoints
