+++ packages/app/package.json
@@ -32,6 +32,12 @@
     "profile:explorer-toggle": "node ./scripts/profile-explorer-toggle.mjs",
     "profile:workspace-switching": "node ./scripts/profile-workspace-switching.mjs",
     "benchmark:agent-stream": "tsx --tsconfig tsconfig.json scripts/benchmark-agent-stream-reducer.ts",
+    "benchmark:draft-attachment-gc": "tsx --tsconfig tsconfig.json scripts/benchmark-draft-attachment-gc.ts",
+    "benchmark:desktop-css-audit": "node scripts/audit-desktop-css-interactions.mjs",
+    "benchmark:desktop-css-interactions": "cross-env E2E_DESKTOP_RUNTIME=1 PASEO_DESKTOP_BENCHMARK=1 playwright test --project='Desktop Chrome' e2e/desktop-css-interactions.benchmark.spec.ts",
+    "benchmark:desktop-interaction": "cross-env E2E_DESKTOP_RUNTIME=1 PASEO_DESKTOP_BENCHMARK=1 playwright test --project='Desktop Chrome' e2e/desktop-interaction.benchmark.spec.ts",
+    "benchmark:desktop-streaming": "cross-env E2E_DESKTOP_RUNTIME=1 PASEO_DESKTOP_BENCHMARK=1 playwright test --project='Desktop Chrome' e2e/desktop-streaming.benchmark.spec.ts",
+    "benchmark:desktop-markdown": "cross-env E2E_DESKTOP_RUNTIME=1 PASEO_DESKTOP_BENCHMARK=1 PASEO_MARKDOWN_BENCHMARK=1 playwright test --project='Desktop Chrome' e2e/desktop-streaming.benchmark.spec.ts",
     "profile:workspace-tabs": "node ./scripts/profile-workspace-tabs.mjs",
     "deploy:web": "npm run build:web && wrangler pages deploy dist --project-name paseo-app --branch main",
     "build:terminal-webview": "node ./scripts/build-terminal-webview-html.mjs",
