+++ packages/app/playwright.config.ts
@@ -31,7 +31,10 @@ export default defineConfig({
   projects: [
     {
       name: "browser",
-      testIgnore: ["**/*.real.spec.ts"],
+      testIgnore: [
+        "**/*.real.spec.ts",
+        ...(process.env.PASEO_DESKTOP_BENCHMARK === "1" ? [] : ["**/*.benchmark.spec.ts"]),
+      ],
       use: { ...devices["Desktop Chrome"] },
     },
     {
