+++ packages/app/scripts/benchmark-draft-attachment-gc.ts
@@ -1,7 +1,6 @@
 import { writeFileSync } from "node:fs";
 import { performance } from "node:perf_hooks";
-import { summarizeSamples } from "../../../scripts/benchmarks/stats";
-import type { BenchmarkTaskResult } from "../../../scripts/benchmarks/types";
+import { summarizeSamples, type BenchmarkTaskResult } from "./benchmark-support";
 import {
   collectStreamUserImageIds,
   haveDraftImageReferencesChanged,
