+++ packages/app/src/stores/draft-store/state.ts
@@ -8,7 +8,7 @@ import type { StreamItem } from "@/types/stream";
 import { z } from "zod";
 
 export const DRAFT_STORE_VERSION = 5;
-export const FINALIZED_DRAFT_TTL_MS = 5 * 60 * 1000;
+const FINALIZED_DRAFT_TTL_MS = 5 * 60 * 1000;
 
 export interface LegacyDraftImage {
   uri: string;
