+++ packages/app/src/utils/markdown-block-window.ts
@@ -1,8 +1,7 @@
 export const MAX_MOUNTED_MARKDOWN_BLOCKS = 96;
 export const MOUNTED_MARKDOWN_HEAD_BLOCKS = 32;
 export const MAX_PROMOTED_ASSISTANT_MARKDOWN_BLOCKS = MOUNTED_MARKDOWN_HEAD_BLOCKS;
-export const MOUNTED_MARKDOWN_TAIL_BLOCKS =
-  MAX_MOUNTED_MARKDOWN_BLOCKS - MOUNTED_MARKDOWN_HEAD_BLOCKS;
+const MOUNTED_MARKDOWN_TAIL_BLOCKS = MAX_MOUNTED_MARKDOWN_BLOCKS - MOUNTED_MARKDOWN_HEAD_BLOCKS;
 
 export interface MarkdownBlockWindow<T> {
   items: readonly T[];
